+++ Corrstown/MHSystems/app/src/main/java/com/mh/systems/corrstown/activites/ClubNewsDetailActivity.java
@@ -310,7 +310,7 @@ public class ClubNewsDetailActivity extends BaseActivity {
      */
     public void showAlertOk(String strMessage) {
 
-        if (builder == null) {
+        if (builder == null && !isFinishing()) {
             builder = new AlertDialog.Builder(this);
             builder.setMessage(strMessage)
                     .setCancelable(false)