+++ Sunningdale/MHSystems/app/src/main/java/com/mh/systems/sunningdale/ui/activites/TopUpActivity.java
@@ -141,8 +141,8 @@ public class TopUpActivity extends BaseActivity {
         //Get Closing balance.
         strClosingBalance = getIntent().getExtras().getString("strClosingBalance");
         if (strClosingBalance.length() > 0) {
-            tvYourBalance.setText((getString(R.string.text_title_your_balance)
-                    + " " + strClosingBalance));
+            tvYourBalance.setText(decimalFormat.format((getString(R.string.text_title_your_balance)
+                    + " " + strClosingBalance)));
 
             fCardBalance = Float.parseFloat(strClosingBalance.substring(1, strClosingBalance.length()));
         }
@@ -438,7 +438,6 @@ public class TopUpActivity extends BaseActivity {
                 strMaxTopup = topUpPriceListResponse.getData().getMaxTopupStr();
 
                 tvCurrencySign.setText(topUpPriceListResponse.getData().getCrnSym());
-
 
                 if (topUpPriceListResponse.getData().getTopupList().size() > 0) {